logging
git-svn-id: file:///home/tobias/.svn/repos/trunk@13 dab74e7a-dce9-0310-a56b-ef9c8ba74961

+++ src/de/gliderpilot/preferences/DimensionPreference.java
@@ -75,7 +75,7 @@ public class DimensionPreference extends AbstractPreference {
 	}
 		
 	public void setDimension(Dimension dimension) {
-		this.dimension.setSize(dimension.getWidth(), dimension.getHeight());
+		this.dimension.setSize((int) dimension.getWidth(), (int) dimension.getHeight());
 	}
 
 	class DimensionPrefPanel extends CancelApplyPanel implements Prefs.Listener {
